後台-強還：增加機車安心服務計算

+++ Domain/SP/Output/Car/SPOutput_GetCarMillage.cs
@@ -7,6 +7,6 @@ namespace Domain.SP.Output.Car
         /// <summary>
         /// 里程數
         /// </summary>
-        public int Millage { get; set; }
+        public float Millage { get; set; }
     }
 }
